remove as i64

+++ crates/sdk/src/extensions.rs
@@ -235,7 +235,7 @@ impl ScalarTypeExt for ScalarType {
         let v = quote!(#pcr::PrismaValue);
 
         match self {
-            ScalarType::Int => quote!(#v::Int(#var as i64)),
+            ScalarType::Int => quote!(#v::Int(#var)),
             ScalarType::BigInt => quote!(#v::BigInt(#var)),
             ScalarType::Float => quote!(#v::Float(#var)),
             ScalarType::Decimal => quote!(#v::String(#var.to_string())),
